Revert "v38: Move Banner to main export (#6720)"
This reverts commit 0466819e4573155a347847bb6850d46c03525d41.
