--- .changeset/fast-countries-remain.md
@@ -1,5 +1,0 @@
----
-"@primer/react": major
----
-
-v38: Move Banner to main export
